Move ConversationDataSource to another package
